# HW2-MarketplaceAPI/catalog-service/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: "Marketplace API"
  version: "1.0.0"
  description: "API для управления товарами и заказами маркетплейса"
servers:
  - url: "http://localhost:8000"
    description: "Local development"
paths:
  /auth/register:
    post:
      operationId: register
      summary: "Регистрация пользователя"
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: "Успешная регистрация"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
  /auth/login:
    post:
      operationId: login
      summary: "Аутентификация"
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: "Access и refresh token"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/TokenInvalid'
  /auth/refresh:
    post:
      operationId: refresh
      summary: "Обновление access token"
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: "Новые токены"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/RefreshTokenInvalid'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    ValidationError:
      description: "Ошибка валидации"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: VALIDATION_ERROR
            message: "Ошибка валидации входных данных"
            details:
              errors:
                - field: "name"
                  message: "Field required"
    TokenInvalid:
      description: "Невалидный или просроченный токен"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            expired:
              value:
                error_code: TOKEN_EXPIRED
                message: "Токен просрочен"
                details: null
            invalid:
              value:
                error_code: TOKEN_INVALID
                message: "Невалидный токен"
                details: null
    RefreshTokenInvalid:
      description: "Невалидный refresh token"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: REFRESH_TOKEN_INVALID
            message: "Невалидный или просроченный refresh token"
            details: null
...
  schemas:
    UserRole:
      type: string
      enum: [USER, SELLER, ADMIN]
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          $ref: '#/components/schemas/UserRole'
          default: USER
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    TokenResponse:
      type: object
      required: [access_token, refresh_token, token_type]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: bearer
...
    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true